Updated documentation comments, where some object members were missing.
Changed the locals() method to accept either an array of parameter names, or a space- or comma-separated string giving a list of parameter names.

The port() method was not implemented (!)

Added the queryLocal() method.

Redefined the semantics of href/hrefLocal and query/queryLocal. Now href() and query() return a value NOT including the local parameters. hrefLocal() and queryLocal() will return a value including local parameters

Removed extraneous '=' sign to anchors when used as query string.